improve: logo assets
