+++ packages/logo/package.json
@@ -30,7 +30,7 @@
     "/dist"
   ],
   "scripts": {
-    "build": "run-s .:build:clean .:build:esm .:build:cjs",
+    "build": "run-s .:build:clean .:build:esm .:build:cjs && build-logo",
     ".:build:clean": "rimraf dist",
     ".:build:esm": "tsc -p tsconfig-esm.json",
     ".:build:cjs": "tsc -p tsconfig-cjs.json",
@@ -38,7 +38,8 @@
     "lint-and-fix": "lint-and-fix",
     "lint-staged": "lint-staged",
     "test": "jest --runInBand",
-    "docs": "typedoc"
+    "docs": "typedoc",
+    "generate-assets": "build-logo"
   },
   "devDependencies": {
     "@rocket.chat/eslint-config-alt": "workspace:packages/eslint-config-alt",
@@ -47,6 +48,7 @@
     "@types/jest": "^27.0.2",
     "@types/react": "^17.0.34",
     "@types/react-dom": "^17.0.11",
+    "build-logo": "workspace:tools/build-logo",
     "eslint": "^8.2.0",
     "jest": "^27.3.1",
     "lint-all": "workspace:tools/lint-all",
