+++ packages/logo/src/RocketChatLogo/RocketChatLogo.spec.tsx
