+++ packages/logo/src/index.ts
@@ -1,1 +1,2 @@
 export { default as RocketChatLogo } from './RocketChatLogo';
+export { default as TaggedRocketChatLogo } from './TaggedRocketChatLogo';
