+++ packages/onboarding-ui/src/common/OnboardingLogo.tsx
@@ -1,13 +1,15 @@
 import { Box } from '@rocket.chat/fuselage';
-import { RocketChatLogo } from '@rocket.chat/logo';
+import { TaggedRocketChatLogo, RocketChatLogo } from '@rocket.chat/logo';
 import type { ReactElement } from 'react';
 
-export const OnboardingLogo = ({ tag }: { tag?: string }): ReactElement => (
+export const OnboardingLogo = (): ReactElement => (
   <Box width='100%' maxWidth={224}>
-    <RocketChatLogo tagTitle={tag} />
+    <RocketChatLogo />
   </Box>
 );
 
 export const OnboardingLogoCloud = (): ReactElement => (
-  <OnboardingLogo tag='CLOUD' />
+  <Box width='100%' maxWidth={224}>
+    <TaggedRocketChatLogo tagTitle='CLOUD' />
+  </Box>
 );
