+++ tools/build-logo/.eslintrc.json
@@ -1,0 +1,6 @@
+{
+  "extends": "@rocket.chat/eslint-config-alt/typescript",
+  "env": {
+    "node": true
+  }
+}
