+++ tools/build-logo/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "build-logo",
+  "private": true,
+  "bin": "./index.mjs",
+  "version": "0.30.0",
+  "dependencies": {
+    "prettier": "^2.5.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "sharp": "^0.29.3"
+  },
+  "devDependencies": {
+    "tools-utils": "workspace:tools/tools-utils"
+  }
+}
